Address ocr review on PR composer fixes (Refs #20)
- Strengthen caret SGR render test: compare baseline (no composer)
  vs composer render and assert the composer adds background/reverse
  styling, rather than a vacuous count >= 1.
- Assert wrapped comment tail text is present in the wrapped-body test.
- Guard remap_cursor against content_width == 0 division.
- Simplify line_start_byte tail expression.
- Make cursor up/down viewport test assert directional scroll changes
  instead of discarding state.
- Extract shared PR composer test fixtures into prs_test_fixtures to
  remove duplicate helpers across the cursor-arrow and composer-focus
  test modules.
